[Refactor] HttpMethod 디렉토리 변경
- enum을 constant 디렉토리에서 관리
