+++ src/main/java/constant/HttpMethod.java
@@ -1,4 +1,4 @@
-package model;
+package constant;
 
 public enum HttpMethod {
     GET, POST, PUT, DELETE
