+++ src/main/java/model/HttpRequest.java
@@ -1,5 +1,7 @@
 package model;
 
+import constant.HttpMethod;
+
 import java.util.HashMap;
 import java.util.Map;
