+++ src/main/java/webserver/RequestHandler.java
@@ -7,7 +7,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 import controller.FrontController;
-import model.HttpMethod;
+import constant.HttpMethod;
 import model.HttpRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
